Intelligent Joystick
The new joystick adapts to the present AR perspective. It keeps the view relative to the viewer so whenever he/she moves around the movement will always be fixed on the device´s screen.

Resized some prefabs to better suit the AR scene.
Vuforia AR has been enabled for this test.